refactor(structure): :truck: Renames Core/Allocators Folder into Core/Memory to fit more datastructures inside of it
