+++ tests/unit/Memory/CMakeLists.txt
@@ -1,17 +1,17 @@
-add_executable(ASGE_AllocatorTest
+add_executable(ASGE_MemoryTest
     LinearAllocatorTests.cpp
     ArenaAllocatorTests.cpp
     PoolAllocatorTests.cpp
     StackAllocatorTests.cpp
 )
 
-target_link_libraries(ASGE_AllocatorTest
+target_link_libraries(ASGE_MemoryTest
     PRIVATE
         ASGE
         GTest::gtest_main
 )
 
-target_compile_options(ASGE_AllocatorTest
+target_compile_options(ASGE_MemoryTest
     PRIVATE
         $<$<OR:$<CXX_COMPILER_ID:GNU>,$<CXX_COMPILER_ID:Clang>>:
             -Wall -Wextra -Wpedantic
@@ -20,12 +20,12 @@ target_compile_options(ASGE_AllocatorTest
 )
 
 if(WIN32)
-    add_custom_command(TARGET ASGE_AllocatorTest POST_BUILD
+    add_custom_command(TARGET ASGE_MemoryTest POST_BUILD
         COMMAND "${CMAKE_COMMAND}" -E copy_if_different
             $<TARGET_FILE:SDL3::SDL3>
-            $<TARGET_FILE_DIR:ASGE_AllocatorTest>
+            $<TARGET_FILE_DIR:ASGE_MemoryTest>
     )
 endif()
 
 include(GoogleTest)
-gtest_discover_tests(ASGE_AllocatorTest)
+gtest_discover_tests(ASGE_MemoryTest)
